+++ include/votca/tools/calculator.h
@@ -101,7 +101,10 @@ class Calculator {
   Property LoadDefaultsAndUpdateWithUserOptions(const std::string package,
                                                 const Property &user_options) {
     Property defaults = LoadDefaults(package);
-    UpdateWithUserOptions(defaults, user_options);
+    InjectDefaultsAsValues(defaults);
+    Property user_options_with_defaults = user_options;
+    InjectDefaultsAsValues(user_options_with_defaults);
+    UpdateWithUserOptions(defaults, user_options_with_defaults);
     RecursivelyCheckOptions(defaults);
     return defaults;
   }
@@ -111,6 +114,8 @@ class Calculator {
   bool _maverick;
 
   void OverwriteDefaultsWithUserInput(const Property &p, Property &defaults);
+  // Copy the defaults into the value
+  static void InjectDefaultsAsValues(Property &defaults);
   static void RecursivelyCheckOptions(const Property &prop);
   static bool IsValidOption(const Property &p,
                             const std::vector<std::string> &choices);